+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.3.0] - 30/12/2020.
+* Added Property showCloseIcon, to control if close icon appear.
+* Added Property closeIcon, to build a custom close icon.
+* Added QUESTION dialogType support.
+* Added INFO dialogType to show info mark.
 ## [1.2.1] - 11/10/2020.
 * Added Property buttonsBorderRadius.
 * Safe Call operator on Navigation.of(context) before trying to dissmiss the Dialog.